Extract loading teacher layout, (fix) several bugs
